这个是调整了一下下Grid图的问题 但是它架构设计应该是错的 没有分离 我先乱塞点代码进去让它跑起来

+++ Classes/define.h
@@ -2,8 +2,8 @@
 
 namespace TETRIS
 {
-    const int BLOCK_SIZE = 25.0F;
-    const int BLOCK_HALF = BLOCK_SIZE / 2.F;
+    const float BLOCK_SIZE = 25.0F;
+    const float BLOCK_HALF = BLOCK_SIZE / 2.F;
     const int DEGREE = -90.0F;
 
     const int MAX_COL = 20;
